requirements.txt
annotated-types==0.5.0
anyio==3.7.1
boto3>=1.18.2,<2.0.0
botocore==1.31.60
cffi==1.16.0
click==8.1.7
cryptography==41.0.4
exceptiongroup==1.1.3
fastapi==0.103.2
h11==0.14.0
httptools==0.6.0
idna==3.4
jmespath==1.0.1
# jwt==1.3.1
PyJWT==2.8.0

mysql-connector==2.2.9
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
pyopenssl>=20.0.0
s3transfer==0.7.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==1.26.17
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.20.0
websockets==11.0.3
python-multipart==0.0.6